Split method implementations to their respective header files
